[Fix #291] Fix command to run migrations
Fixes #291
Status: Finished

Summary

This fixes the code in main to run migrations. The issue was that the original
code was using `:jdbc-database-url` vs `:database-url`. It now works for both
migrations and rollbacks. I've also simplified the components started by mount
as we only need to start env and not the db.