fix #432 load more activities
Loading to few activities can cause claims not being displayed as claims
and bounties are loaded via separate XHR requests. This change
implements a quick fix around that making old activities show up.

Long term we should move claims into the response of `/api/open-bounties`
so that they always are available for bounties shown in the UI.

This is also somewhat related to better API pagination as described in
https://github.com/status-im/open-bounty/issues/250